Fixed resize problem.
The resize problem occured if one hit ctrl+s to
hide the side pane and the current output held
a string longer than it could display (wrap needed).

If one reshowed the side pane, the main window
was resized. This is due to the strict policy
to never display a scrollbar in the output which
forces the widget to allocate more space for itself.
I don't know if this is a bug (the widget could simply
free some space and wrap) but it's solved for now.