Make scroll-end detection fuzzy
There seem to be inaccuracies when manually scrolling to bottom.
Because of this, make the bottom detection fuzzy by taking the
current position and the max. position and computing a percentage.
If the two values match >= 99% then we're at the bottom.

I would rather have a proper pixel-perfect solution but I'm
too annoyed by this bug for too long.